+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/AbstractModInitializer.java
@@ -19,6 +19,7 @@ import com.refinedmods.refinedstorage.common.autocrafting.PatternGridData;
 import com.refinedmods.refinedstorage.common.autocrafting.PatternItem;
 import com.refinedmods.refinedstorage.common.autocrafting.PatternState;
 import com.refinedmods.refinedstorage.common.autocrafting.ProcessingPatternState;
+import com.refinedmods.refinedstorage.common.autocrafting.StonecutterPatternState;
 import com.refinedmods.refinedstorage.common.configurationcard.ConfigurationCardItem;
 import com.refinedmods.refinedstorage.common.configurationcard.ConfigurationCardState;
 import com.refinedmods.refinedstorage.common.constructordestructor.BlockBreakDestructorStrategyFactory;
@@ -872,6 +873,12 @@ public abstract class AbstractModInitializer {
                 () -> DataComponentType.<ProcessingPatternState>builder()
                     .persistent(ProcessingPatternState.CODEC)
                     .networkSynchronized(ProcessingPatternState.STREAM_CODEC)
+                    .build()));
+        DataComponents.INSTANCE.setStonecutterPatternState(
+            callback.register(createIdentifier("stonecutter_pattern_state"),
+                () -> DataComponentType.<StonecutterPatternState>builder()
+                    .persistent(StonecutterPatternState.CODEC)
+                    .networkSynchronized(StonecutterPatternState.STREAM_CODEC)
                     .build()));
     }