+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/autocrafting/PatternBakedModel.java
@@ -13,9 +13,16 @@ class PatternBakedModel extends BakedModelWrapper<BakedModel> {
     PatternBakedModel(final ModelBaker modelBaker,
                       final BakedModel emptyModel,
                       final BakedModel craftingModel,
-                      final BakedModel processingModel) {
+                      final BakedModel processingModel,
+                      final BakedModel stonecutterModel) {
         super(emptyModel);
-        this.itemOverrides = new PatternItemOverrides(modelBaker, emptyModel, craftingModel, processingModel);
+        this.itemOverrides = new PatternItemOverrides(
+            modelBaker,
+            emptyModel,
+            craftingModel,
+            processingModel,
+            stonecutterModel
+        );
     }
 
     @Override