+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/autocrafting/PatternUnbakedGeometry.java
@@ -20,6 +20,7 @@ public class PatternUnbakedGeometry implements IUnbakedGeometry<PatternUnbakedGe
     private static final ResourceLocation EMPTY_MODEL = createIdentifier("item/pattern/empty");
     private static final ResourceLocation CRAFTING_MODEL = createIdentifier("item/pattern/crafting");
     private static final ResourceLocation PROCESSING_MODEL = createIdentifier("item/pattern/processing");
+    private static final ResourceLocation STONECUTTER_MODEL = createIdentifier("item/pattern/stonecutter");
 
     @Override
     public BakedModel bake(final IGeometryBakingContext ctx,
@@ -31,7 +32,8 @@ public class PatternUnbakedGeometry implements IUnbakedGeometry<PatternUnbakedGe
             modelBaker,
             requireNonNull(modelBaker.bake(EMPTY_MODEL, modelState, function)),
             requireNonNull(modelBaker.bake(CRAFTING_MODEL, modelState, function)),
-            requireNonNull(modelBaker.bake(PROCESSING_MODEL, modelState, function))
+            requireNonNull(modelBaker.bake(PROCESSING_MODEL, modelState, function)),
+            requireNonNull(modelBaker.bake(STONECUTTER_MODEL, modelState, function))
         );
     }
 
@@ -41,5 +43,6 @@ public class PatternUnbakedGeometry implements IUnbakedGeometry<PatternUnbakedGe
         modelGetter.apply(EMPTY_MODEL).resolveParents(modelGetter);
         modelGetter.apply(CRAFTING_MODEL).resolveParents(modelGetter);
         modelGetter.apply(PROCESSING_MODEL).resolveParents(modelGetter);
+        modelGetter.apply(STONECUTTER_MODEL).resolveParents(modelGetter);
     }
 }